feat: 结束锚点系统 + 初步检测面板 + 自动分析优化
- 新增 Chapter.end_anchor 字段及 fill/check/anchor-score API
- 本地锚点合规检测引擎（jieba + embedding，A/B/A+B 策略）
- 编辑页底部初步检测面板（策略/阈值/分数/持久化）
- 修复生成后自动分析开关逻辑：取消勾选不再触发LLM
- 字元比双变量分离（output_token_multiplier / prompt_char_token_ratio）
- Analysis cancelled 状态前端全生命周期处理
- 章节阅读页紧凑布局 + 前后章节预览 + 锚点显示
- docs/FORK_CHANGES.md 更新

+++ backend/app/models/foreshadow.py
@@ -60,8 +60,8 @@ class Foreshadow(Base):
     
     # === 重要性和优先级 ===
     importance = Column(Float, default=0.5, comment="重要性评分 0.0-1.0")
-    strength = Column(Integer, default=5, comment="伏笔强度 1-10(影响读者多强烈)")
-    subtlety = Column(Integer, default=5, comment="隐藏度 1-10(越高越隐蔽)")
+    strength = Column(Float, default=5.0, comment="伏笔强度 1-10(影响读者多强烈)")
+    subtlety = Column(Float, default=5.0, comment="隐藏度 1-10(越高越隐蔽)")
     urgency = Column(Integer, default=0, comment="紧急度: 0=不紧急, 1=需关注, 2=急需回收")
     
     # === 关联信息 ===
